refactor: FramePins 에서 Strike 객체 분리
