+++ README.md
@@ -59,9 +59,12 @@
     - 2번 투구
 - [X] 두 Pins 의 합은 10 개를 초과할 수 없다.
 - [X] FramePins 생성은 스트라이크인 경우와 그 외의 경우로 나누어진다.
-    - 스트라이크인 경우, 한 번의 투구만 한다.
     - 그 외의 경우, 두 점수 모두 파라미터로 받는다.
 - [X] 파라미터 Null 유효성 검증
+
+### Strike : 스트라이크 핀
+- [X] FramePins 를 상속받는다.
+- [X] 스트라이크는 한 번의 투구만 한다.
 
 ### InputView
 - [X] 플레이어를 입력받아 그대로 리턴한다.
